refactor: . Fix Task/Task.ts > lib/PriorityTools.ts cyclic dependency
