+++ src/Query/Filter/PriorityField.ts
@@ -1,7 +1,8 @@
-import { Priority, Task } from '../../Task/Task';
+import type { Task } from '../../Task/Task';
 import { Explanation } from '../Explain/Explanation';
 import type { Comparator } from '../Sort/Sorter';
 import type { GrouperFunction } from '../Group/Grouper';
+import { Priority } from '../../Task/Priority';
 import { Field } from './Field';
 import { Filter } from './Filter';
 import { FilterOrErrorMessage } from './FilterOrErrorMessage';
