+++ src/TaskSerializer/DefaultTaskSerializer.ts
@@ -1,7 +1,8 @@
 import type { Moment } from 'moment';
 import { type TaskLayoutComponent, TaskLayoutOptions } from '../Layout/TaskLayoutOptions';
 import { Recurrence } from '../Task/Recurrence';
-import { Priority, Task, TaskRegularExpressions } from '../Task/Task';
+import { Task, TaskRegularExpressions } from '../Task/Task';
+import { Priority } from '../Task/Priority';
 import type { TaskDetails, TaskSerializer } from '.';
 
 /* Interface describing the symbols that {@link DefaultTaskSerializer}
