+++ tests/Commands/CreateOrEditTaskParser.test.ts
@@ -3,9 +3,9 @@
  */
 import moment from 'moment';
 import { resetSettings, updateSettings } from '../../src/Config/Settings';
-import { Priority } from '../../src/Task/Task';
 import { taskFromLine } from '../../src/Commands/CreateOrEditTaskParser';
 import { GlobalFilter } from '../../src/Config/GlobalFilter';
+import { Priority } from '../../src/Task/Priority';
 
 window.moment = moment;
