+++ tests/DocumentationSamples/UrgencyDocs/DocsSamplesForUrgency.test.ts
@@ -7,7 +7,8 @@ import moment from 'moment';
 import { verifyMarkdownForDocs } from '../../TestingTools/VerifyMarkdown';
 import { TaskBuilder } from '../../TestingTools/TaskBuilder';
 import { Urgency } from '../../../src/Task/Urgency';
-import { Priority } from '../../../src/Task/Task';
+
+import { Priority } from '../../../src/Task/Priority';
 
 window.moment = moment;
