+++ tests/Query/Filter/FunctionField.test.ts
@@ -6,7 +6,6 @@ import moment from 'moment';
 import { FunctionField } from '../../../src/Query/Filter/FunctionField';
 import { Status } from '../../../src/Statuses/Status';
 import type { Task } from '../../../src/Task/Task';
-import { Priority } from '../../../src/Task/Task';
 import {
     toGroupTaskFromBuilder,
     toGroupTaskUsingSearchInfo,
@@ -22,6 +21,7 @@ import {
 import { fromLine } from '../../TestingTools/TestHelpers';
 import { Query } from '../../../src/Query/Query';
 import { TasksDate } from '../../../src/Scripting/TasksDate';
+import { Priority } from '../../../src/Task/Priority';
 
 window.moment = moment;
