+++ tests/Task/Urgency.test.ts
@@ -5,9 +5,9 @@
 import moment from 'moment';
 import { TaskBuilder } from '../TestingTools/TaskBuilder';
 import { Urgency } from '../../src/Task/Urgency';
-import { Priority } from '../../src/Task/Task';
 import { calculateRelativeDate } from '../TestingTools/DateTestHelpers';
 import { fromLine } from '../TestingTools/TestHelpers';
+import { Priority } from '../../src/Task/Priority';
 
 window.moment = moment;
